resume: split components and add transition
