+++ package.json
@@ -20,6 +20,7 @@
     "@mui/material-nextjs": "^5.16.4",
     "next": "^14.2.5",
     "react": "^18.3.1",
+    "react-awesome-reveal": "^4.2.14",
     "react-dom": "^18.3.1",
     "swiper": "^11.1.8",
     "swr": "^2.2.5"
